stt-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
boto3==1.28.85
pyyaml==6.0.1
openai-whisper==20231117
ffmpeg-python==0.2.0
